Moved habit tests to internal/habit with testdata
